Remove the delayed-hide of previous content state
This effectively reverts the intent of b9503be6.

The concept of delaying the hide of the previous content state is
causing more trouble than it is worth, resulting in several race
conditions.  Rather than resolve the race conditions, it is more
prudent to remove the delayed hide entirely, as there is no reason
to think that atomically swapping visibility would be broken.