Allow MSV to delegate whether its child can scroll vertically
This is especially useful for widgets such as SwipeRefreshLayout, which
enables its functionality based on whether the child can scroll or not.
By default, MSV is not scrollable, which causes SRL to think it should
always be able to activate the gesture.  Instead, if the child is scrollable,
we should pass that along to the caller, so that the swipe gesture can
be properly disabled when the content is still scrollable.